Final version

+++ src/include/zombies.h
@@ -23,8 +23,8 @@ typedef struct zombie_list {
 } zombie_list;
 
 typedef struct zombie {
-    int x;
-    int y;
+    float x;
+    float y;
     int health;
     SDL_Texture *texture;
     SDL_Rect rect;
@@ -32,13 +32,21 @@ typedef struct zombie {
 } zombie;
 
 // Zombie Functions
+// Create Zombie
 zombie* create_zombie( SDL_Texture *texture);
+// Move zombie towards the player
 void move_zombie(zombie* z, player* p , int score , float difficulty);
+// Add zombie to zombie list
 void add_zombie(zombie_list** zl, zombie* z);
+// Remove zombie from zombie list
 void remove_zombie(zombie_list** zl, zombie* z) ;
+// Move all zombies towards the player using move_zombie function
 void move_zombies(zombie_list** zl , player* p ,int score , float difficulty);
+// Free zombie list
 void free_zombie_list(zombie_list** zl);
+// Make zombie face the player
 void rotate_zombie(zombie* z , player* p) ;
+// Generate zombies based on a spawn rate as well as score
 void zombie_generator(zombie_list** zl , SDL_Texture *texture , int p , int score , float difficulty) ;
 
 #endif
